+++ src/coalesce-vue-vuetify3/src/components/admin/c-admin-audit-log-page.vue
@@ -450,6 +450,9 @@ defineExpose({
   }
   td {
     vertical-align: top;
+    pre {
+      margin: 0;
+    }
   }
 }